+++ src/main/java/cc/polyfrost/oneconfig/renderer/scissor/Scissor.java
@@ -25,7 +25,7 @@ public class Scissor {
         this.height = scissor.height;
     }
 
-    public boolean isInScissor(int x, int y) {
+    public boolean isInScissor(float x, float y) {
         return x >= this.x && x <= this.x + this.width && y >= this.y && y <= this.y + this.height;
     }
 }
